Migrate from CircleCI to GitHub Actions (#51)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pplbench/pull/51

GitHub Actions seems to be recommended over CircleCI for running automatic workflows. Additionally, because we already got a GitHub Actions workflow for building the website, it'd be nice to migrate all of the workflows to the same platform.

A few notes about this commit:

- Previously, our GitHub page is only built when a change is pushed to master. As a result, there were a few cases where we only found out the build error after a change is merged. So I added pull request to the workflow trigger condition so that build error will be visible from pull request.
- Instead of manually set the number of parallel jobs using the `parallism` key in CircleCI, we can now programmatically create new jobs using the output of a previous job. This means that we can dynamically look for JSON files in PPL Bench and create a job for each of them
  - In fact, we can even run a combination of a list of Python versions, operating systems, and JSON files in parallel. The current config will only create jobs on Python 3.6 and Python 3.7 for each of the JSONs on Ubuntu 18.04 (a.k.a, `ubuntu-latest`), but we can add more configs in the future if needed.
- GitHub Actions has a marketplace and make it easy to perform a wider ranges of actions -- for example, instead of manually create a release after bumping the version of PPL Bench, it is now possible to automatically publish a new release on GitHub and upload the package to PyPI when we bump the version in `pplbench/__init__.py`.

We'll need to turn off CircleCI before this commit is merged, so that it's not going to complain about missing config file.

I'm also removing the CircleCI status badge from README for now. This should be replaced by the status badge for GitHub Actions after the config is merged to master (otherwise we can't really tell if the badge is working or not).

Reviewed By: nimar

Differential Revision: D24578372

fbshipit-source-id: 169e9d20df0964561f813f76c18304b0d6d5c8bc