Rename num_samples to iterations (#57)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pplbench/pull/57

As per the discussion in D24580278, the term "num_samples" does sound a bit confusing as people might expect it to equal the number of post-warmup samples. Therefore, in this diff, the config is renamed to "iterations" to avoid confusion.

(The reason "iterations" is used instead of "iter" is that the latter collides with Python's built-in function  `iter()`)

Reviewed By: neerajprad, nimar

Differential Revision: D24866712

fbshipit-source-id: 5c6759ba63abb64edaa56b0424f720791615af2c