Enable License Lint for Bean Machine and PPL Bench (#96)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pplbench/pull/96

Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1316

This diff enable license lint for Bean Machine and PPL Bench, which will enforce a copyright statement on top of every file under PPL Team's codebase. This will also give us a diff-time warning so we can avoid accidentally committing codes without the license and failing OSS requirement (such as T110683359).

I was hoping to enable this linter for a while so we don't have to keeping track of the headers manually, but license lint was using the "Facebook" variant of the headers. Now that it's updated to use "Meta," we can safely enable it for our project without having to do a major codemod later on :).

In addition to enabling the linter, this diff also updates the few files that failed the license requirement. These have to be done in the same diff to make the linter happy.

Reviewed By: neerajprad

Differential Revision: D33772236

fbshipit-source-id: 662555e81527fdb55b9279c729fecbc072ba6a05