Changes to be compatible with jsonargparse v4 (#92)
Summary:
### Motivation
I have noticed that this project has as dependency jsonargparse and makes use of the `ActionJsonSchema` class. Scheduled for tomorrow (Nov 16th 2021) jsonargparse v4.0.0 will be released and it includes a breaking change in the behavior of `ActionJsonSchema`. The change is that the values parsed with this action will no longer be a nested namespace but a nested dict instead.

### Changes proposed
This pull request includes a small change so that there is compatibility with the newer versions of jsonargparse.

Pull Request resolved: https://github.com/facebookresearch/pplbench/pull/92

Test Plan:
I tested by using `pplbench examples/example.json` with jsonargparse v4.0.0rc1 making sure that it worked just like with previous versions.

### Types of changes
- [x] Docs change / refactoring / dependency upgrade
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

### Checklist
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **[CONTRIBUTING](https://github.com/facebookresearch/pplbench/blob/main/CONTRIBUTING.md)** document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.
- [x] The title of my pull request is a short description of the requested changes.

Reviewed By: jpchen

Differential Revision: D32424133

Pulled By: horizon-blue

fbshipit-source-id: 20500fadbe41206169d57f8ed9f4925fcd5fba82