Fix lint issue & add Bean Machine to example jsons (#100)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pplbench/pull/100

Our CI is having trouble running PPL Bench after we exported the BM models last week, so this diff is meant to address that.

Because BM is public now, there is no need to maintain a separate JSON directory under  `pplbench/ppls/beanmachine/examples`, so the content of those JSON configs are merged to the root directory.

Differential Revision: D34087282

fbshipit-source-id: 2750f7b782af313c7158c57f2fa35fe61d0e523d