PPLBench Profiling Schema and an Example JSON file
Summary:
I added profiling support to PPLBench, so that it serves as an
infrastructure for performance management. This Diff is the first
of a series to commit this work. There are two profilers included
in the infratructure: The first is deterministic and works by
adding code to the profiled BM code (Python or C++). The second is
based on statistical sampling and uses an external process to do
the sampling (note: this has nothing to do with the sampling work
of the models that are using BM. We are here sampling for
performance). Both profilers have advantages and disadvantages. Eric Lippert wrote the deterministic profiler for BMG (C++
side). It is hoped that in the future, one can drop a model to
the PPLBench infrastructure and perform measurements and profiling
with ease.

This Diff includes the necessary change to the PPLBench
schema in order to enable profiling. The changes include five
non-mandatory parameters, namely:
* profile_run: Boolean. If true, it enables the profiler.
* num_profiled: Integer. This specifies the desired number of
functions to be reported. For example, specifying 10 will
produce a report of the top 10 functions in CPU utilization. This
parameter is relevant only if the profiling is deterministic
(see below); it is ignored for statistical profiling.
* strip_profiled_names: Boolean. If true, it masks the directory
path of the function in the resulting flame graph. This
improves readability of the graph, but can create a problem
if there are functions with the same name. The flamegraph
parser cannot disambiguate functions if there are aliases. If
your code has functions of the same name in different locations,
you should set this flag to false.
* profiling_tool_dir: String. This specifies the path to the third
party profiling tools that were imported into PPLBench. Currently
it is in the lib directory under pplbench.
* profiling_type: String. Either deterministic or statistical.
This flag specifies the type of desired profiling. Note: For
profiling BMG code, the deterministic profiler produces
two sets of output files, one for the Python code and one
for the C++ part. The deterministic profiler works inside
the functions, and therefore is not aware of the runtime
of the other part of the code. The statistical profiler does
not have this problem.

I include a sample JSON file to show how to use the schema.

Reviewed By: rodrigodesalvobraz

Differential Revision: D35830377

fbshipit-source-id: 37a74ed2d005ea853d495ca5160b1e6328c3a529