Improving the Readability of PPLBench Output (#110)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pplbench/pull/110

PPLBench provides a number of interesting files. One of which is the configuration that was tested. This is useful if one is doing many performance runs and wants to disambiguate the various runs from one another later. However, the printed file was not very readable, as it consists of a very long string without line breaks. This modification here fixes this minor problem.

Reviewed By: ericlippert, horizon-blue

Differential Revision: D35833204

fbshipit-source-id: 49738411169763e06e96afb9836439d6c55481cd