Profiler infrastructure for Bean Machine
Summary:
This commit includes all the files that I have added to support two styles of profiling in PPLBench, namely statistical profiling and deterministic profiling. Please see Diff D35830377 (https://github.com/facebookresearch/pplbench/commit/0ce10772f7bd0329ca1318764abd5ef994e7c588) for more details on how it works. The Diff includes the following:
ppl_profiler.py: This is the majority of the profiling support, implemented as a Python class. It gets called from inference.py or from main.py. If the .json file does not specify profiling, this object is silent and causes no overhead.
main.py: I added the interface to the profiler. The file imports the profiler object, and also parses the part of the JSON file configuration input that specifies the parameters for the profiling run.
inference.py: This file interfaces the profiler infrastructure in Python with the profiling infrastructure in BMG. The profiler infrastructure in BMG will be submitted in a subsequent Diff.
py-spy: This is a third-party sampler tool that generates a flamegraph and other statistics. This is open source with an MIT license. It is written in Rust. This file is in a directory labeled third-party under lib. This way, there is conflict between the mandatory inclusion of the third party license and the Meta license.
flamegraph.pl: This is a PERL script that reads the output of the profiler and generates an interactive flame graph. It is a third party tool, under its own fashioned MIT license. Again, this is placed under the third-party directory for the same reasons as py-spy.

Reviewed By: ericlippert

Differential Revision: D36268506

fbshipit-source-id: b2569ae5315e1a39fc64bcbb5dab4b17800b78aa