Bump version to v0.0.3
Summary: During the conversation with Mootaz today, it just came to my realization that we haven't been making new release for PPL Bench for more than a year, and there has been a lot of changes both on our side, and on our dependencies. In so far many of us has been installing PPL Bench from source to get the latest changes, but it'd be more convenient if users can get PPL Bench from PyPI without running into errors.

Differential Revision: D38062970

fbshipit-source-id: 49448628103293ea4a72e157043c4eeaa885cc7f