PPLBench Interface to BMG Statistics (#120)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pplbench/pull/120

A new statistics gathering package about the BMG graph was introduced in Diff 38284547. The current diff contains code for interfacing PPLBench to the statistics package. The modifications are in main.py, to generate the output file containing the statistics, and in inference.py, which contains the code to actuall call the statistics package upon the formation of the graph. The graph does not change during the PPLBench runs, even with different inference methods, so one call is sufficient to produce the results.

Differential Revision: D38477112

fbshipit-source-id: 52a353d0594ea98e9dc95c9ed855d06c8d08c194