JSON & Scripts to Streamline Bean Machine Benchmarking (#121)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pplbench/pull/121

This set of files simplifies the running of the PPLBench benchmarks for BM-Python and BMG implementations. Prior to this diff, one had to manually update the json files to eliminate the runs that are not relevant to a study. The example JSON files are organized by benchmarks, but often a study is focused on a particular inference implementation. This diff provides a more streamlined approach that is oriented by the inference mechanism to simplify the process of running the benchmark and increase the convenience for end user. At some point, the parameters included in the JSON files will need to be tuned to be more approriate for the inference algorithm that is being tested. This work is left for a future iteration. This is a first stake in the ground that should be improved in the future.

Reviewed By: CactusWin

Differential Revision: D38600089

fbshipit-source-id: 9fa1cd57409eee8bba9691f1fd77450b2c1b89e0