Add Bean Machine VI to PPL Bench
Summary:
Adding VI support to PPL Bench so we can easily benchmark the performance of different configurations on the models there.

To draw the samples that PPL Bench expects at the end, here we simply takes the VI distributions at the end of an inference (conditioned on the learned parameters) and draw N samples from them, where N equals number of iterations.

Differential Revision: D40951663

fbshipit-source-id: 1a04fc658c027684811a58c7d5bdf158e231e6f6