remove unsupported/unused benchmarks that cause test warnings in production diffs
Summary:
**For now, this is an RFC (request for comments).**
If you are interested in maintaining the code suggested for deletion here, please step up.

The code in `pplbench` causes warnings in diffs like D48606875. This causes distractions when trying to ship diffs.

 {F1075724799}

Clicking through the logs identifies `pplbench` as the culprit: {F1075724803}

In the interest of keeping a good signal/noise ratio, I suggest to remove this code -- unless someone is willing to step up and maintain it. As far as I can tell, it does not serve any production purpose here at Meta.

=== Source syncing note ===

Starting in February, the source of truth of the Beanmachine open source project was switched from Meta internal to GitHub, ShipIt was deactivated in D43404340.

The `benchmarks/` directory does not show up on [GitHub](https://github.com/facebookresearch/beanmachine), so I assume this part is not synced.

Reviewed By: ToddSmall

Differential Revision: D48647611

fbshipit-source-id: b41d436b283808217b7736cf82d13a76dcc30713